Patrol() and Chasing()
Patrol():

Made sure that objects get removed from the list when they are detected
to no longer be "hot" or "cold".  Enemy then gets a new path to the next
object.

Chasing():

Started on chasing.  Working on how to get the enemy to chase the player
and not create too many unnecessary paths.